Start to include the TwiM Reader for MDPP16
Some debugging to make it compiling

add some comments + I checked that TIME and ENERGY data of MDPP16 are ordered. For each input signal above the threshold, in the data stream there is first the time data word and then the energy data word.

add the TrefMusicMappedData and fill it in the TwimReader

Add a macro to check the TwimMappedData