Bug fix in TPAT reconstructions in main_online.C and R3BSofOnlineSpectra.cxx
The R3BSofSciTcal2RawPos/TofPar classes are now using gaussian fit of the peak for accurate analysis. +/- 5 sigma range is now applied.

Simplified the reconstruction algorithm of R3BSofSciTcal2SingleTcal. Now the S2 SofSci and the S8 FRS scintillator are used equally.
The multiplicity of CaveC start plastic is defined the smaller one of either S2-CaveC or S8-CaveC ToF to recover missing events due to high rate at S2.

Modification in the classes for Twim. For the dedicated analysis getting rid of the beta dependency, it is necessary to get averaged energy.
Similar modification has been applied in the r3bmusic classes, in the R3BRoot repository.